Implement trading
Trading is implemented like so:

A user makes a trade offer to another user, offering 1 or more emojis (which they must have) and requesting 1 or more emojis from the other user (which they don't need to have).

There can only be one trade offer from a specific user to another specific user, but the offering user can withdraw and the targeted user can reject. The targeted user can also still make an offer the other way.

At any time, users can view all their incoming and outgoing offers.

Then, when a user accepts an offer, it first checks whether that offer is valid (still exists, user has the requested emojis, offerer still has the offered emojis). Then the user is told the contents of the trade and asked to confirm. If, by the time the user confirms, the trade became invalid, or if it is different from what it was in any way, the confirmation fails.

If acceptance goes through, the trade is logged, the offer is deleted, and emoji counts are adjusted as appropriate. Finally, any emoji entries with a count of 0 are deleted.