Added recycling
Users can sacrifice 3 emojis to get 1 new random one.
I ended up up conceptualizing this as a sort of trade with the system, where the system's user ID is 0. The "trade" is added to the trade log this way.
Users are guaranteed not to get one of the emojis they put in back, although the chance of that is pretty slim anyway.